// webapp/ui/src/app/api-client/models/bet-participation-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BetParticipationResponse {
  created_at: string;
  id: number;
  lottery_id: number;
  meme_coin_address: string;
  sol_amount: number;
  tx_signature?: (string | null);
  user_id: number;
  wallet_address: string;
}